neufin-backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
supabase
pandas
yfinance
reportlab
anthropic
google-generativeai
openai
groq
stripe
sentry-sdk[fastapi]
pydantic
pydantic-settings
